[cleanup] start splitting the code across multiple files
